+++ src/pages/Mission.vue
@@ -190,31 +190,18 @@
               :class="paramValueClass(props.row) + ' q-mb-md'"
                style="white-space: normal"
             >
-              <div>
-                <parameter-value
-                  :ref="`parameter-value_${props.row.paramName}`"
-                  class="q-pa-xs"
-                  style="font-family:monospace;min-width:24em;word-break:break-all;font-size:0.9em"
-                  :required="props.row.required"
-                  :param-type="props.row.type"
-                  :param-value="props.row.paramValue"
-                />
-
-                <q-popup-edit
-                  :buttons="editable()"
-                  v-model="props.row.paramValue"
-                  @save="saveArguments(props.row)"
-                >
-                  <parameter-value-input
-                    :param-name="props.row.paramName"
-                    :param-required="props.row.required"
-                    v-model="props.row.paramValue"
-                    :param-type="props.row.type"
-                    :default-value="props.row.defaultValue"
-                    :editable="editable()"
-                  />
-                </q-popup-edit>
-              </div>
+              <parameter-value
+                :ref="`parameter-value_${props.row.paramName}`"
+                class="q-pa-xs"
+                style="font-family:monospace;min-width:24em;word-break:break-all;font-size:0.9em"
+                :required="props.row.required"
+                :param-name="props.row.paramName"
+                :param-type="props.row.type"
+                :param-value="props.row.paramValue"
+                :default-value="props.row.defaultValue"
+                :editable="editable()"
+                @save="val => { props.row.paramValue = val; saveArguments(props.row) }"
+              />
             </q-field>
           </q-td>